requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Brotli==1.1.0
cachetools==5.4.0
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
cloudinary==1.40.0
cssselect2==0.7.0
Django==5.0.5
django-cors-headers==4.4.0
django-filter==24.2
django-storages==1.14.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
fonttools==4.54.1
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
gunicorn==22.0.0
html5lib==1.1
idna==3.7
inflection==0.5.1
mysql-connector-python==9.0.0
numpy==2.1.0
openpyxl==3.1.5
packaging==24.0
pandas==2.2.2
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.27.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.3
pydantic_core==2.18.4
pydyf==0.11.0
PyJWT==2.8.0
pyphen==0.16.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
rsa==4.9
six==1.16.0
sqlparse==0.5.0
tinycss2==1.3.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
weasyprint==62.3
webencodings==0.5.1
zopfli==0.2.3
